requirements.txt
streamlit>=1.33.0
python-dotenv>=1.0.1
matplotlib>=3.8.0
google-generativeai>=0.7.0